+++ My Project/AssemblyInfo.vb
@@ -7,5 +7,5 @@ Imports System.Runtime.InteropServices
 <Assembly: AssemblyCopyright("stax76")>
 <Assembly: ComVisible(False)> 
 <Assembly: Guid("54d5dc0f-6ba7-42f0-a82f-9c71d561cb8b")>
-<Assembly: AssemblyVersion("1.6.1.6")>
-<Assembly: AssemblyFileVersion("1.6.1.6")>
+<Assembly: AssemblyVersion("1.6.1.7")>
+<Assembly: AssemblyFileVersion("1.6.1.7")>